nodemon & docker compose upd
