+++ services/api-gateway/.env
@@ -1,5 +1,5 @@
 PORT=3000
 NODE_ENV=development
 
-WEATHER_SERVICE_URL=http://localhost:3001
-SUBSCRIPTION_SERVICE_URL=http://localhost:3002
+WEATHER_SERVICE_URL=http://weather-service:3001
+SUBSCRIPTION_SERVICE_URL=http://subscription-service:3002
