+++ services/api-gateway/package.json
@@ -5,6 +5,7 @@
   "scripts": {
     "build": "rm -rf dist && tsc && tsc-alias",
     "start": "node dist/main.js",
+    "start:dev": "nodemon -L src/index.ts",
     "dev": "tsx src/main.ts"
   },
   "keywords": [],
@@ -12,17 +13,18 @@
   "license": "ISC",
   "description": "",
   "dependencies": {
+    "@fastify/formbody": "^8.0.2",
     "axios": "^1.10.0",
     "fastify": "^5.4.0",
-    "zod": "^4.0.5",
-    "tsconfig-paths": "^4.2.0"
+    "tsconfig-paths": "^4.2.0",
+    "zod": "^4.0.5"
   },
   "devDependencies": {
     "@types/node": "^22.15.18",
-    "tsx": "^4.20.3",
-    "typescript": "^5.8.3",
+    "nodemon": "^3.1.10",
     "ts-node": "^10.9.2",
     "tsc-alias": "^1.8.16",
-    "nodemon": "^3.1.10"
+    "tsx": "^4.20.3",
+    "typescript": "^5.8.3"
   }
 }
