+++ services/api-gateway/src/routes/index.ts
@@ -5,6 +5,6 @@ import { healthRoutes } from "./health";
 
 export async function routes(fastify: FastifyInstance) {
   await fastify.register(weatherRoutes, { prefix: "/api/weather" });
-  await fastify.register(subscriptionRoutes, { prefix: "/api/subscription" });
+  await fastify.register(subscriptionRoutes, { prefix: "/api/" });
   await fastify.register(healthRoutes, { prefix: "/api/health" });
 }
