+++ services/api-gateway/src/routes/subscription/index.ts
@@ -1,11 +1,16 @@
 import { FastifyInstance } from "fastify";
 
 export async function subscriptionRoutes(fastify: FastifyInstance) {
-  fastify.post("/", async (request, reply) => {
-    const { email, city } = request.body as { email: string; city: string };
+  fastify.post("/subscribe", async (request, reply) => {
+    const { email, city, frequency } = request.body as {
+      email: string;
+      city: string;
+      frequency: string;
+    };
     const result = await fastify.subscriptionClient.createSubscription(
       email,
-      city
+      city,
+      frequency
     );
     reply.send(result);
   });
