+++ services/subscription-service/.env
@@ -1,6 +1,6 @@
 PORT=3002
 NODE_ENV=development
-BASE_API_URL=http://localhost:3002
+BASE_API_URL=http://subscription-service:3002
 
 DATABASE_URL=postgresql://postgres:postgres@db:5432/weather
 RESEND_API_KEY=re_8jETgjPt_9j5eJnNWXrYPkyNRc633LAeB
